pyproject.toml
[project]
name = "data-task"
version = "0.1.0"
description = "RMT Data Engineering Task"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "psycopg[binary]>=3.2.6",
    "psycopg>=3.2.0",
    "clickhouse-driver>=0.2.0",
    "ruff>=0.11.0",
    "apache-airflow==2.10.3",
    "apache-airflow-providers-postgres>=6.1.3",
    "airflow-clickhouse-plugin>=1.4.0"
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
